Overhaul Lua-shader system to use wrapper classes
This is to fix an issue where Lua-shaders would not get destroyed on level change because shaders have to be persistent, but they would *have* to be destroyed because they depend on the Lua state. The new system seprates the base shader (which can persist) and the Lua representation of the shader (which can be destroyed on level change).