fix: temporarily disable multi-threading for animations
panima-animations can affect arbitrary component properties, which cannot be guaranteed to be thread-safe.
To provide some degree of multi-threading, the animation values could be computed on multiple threads and stored in temporary storage, then applied on the main thread.
Additionally, individual properties could be marked as thread-safe, which could then be applied on the secondary threads immediately.